added break statement
Two important notes:

1. using throw and catch in the Interpreter is better then setting a
flag, as with an exception it is possible to only change the visitWhile
function. Otherwise it would be necessary to check the flag in the
visitBlock function.
2. continue is not that easy to implement. To make this work, the
desugaring of the for loop must be changed, otherwise if continue is
used inside 'for' the increment part is executed. This should be
possible with 'while' and 'break' but the better way would be
translating for directly in this case, I guess.